+++ Point Cloud Processor/CustomAVLTree.h
@@ -1,13 +1,13 @@
 #pragma once
 struct LinkedListNode {
-	double x;
-	double y;
-	double z;
+	float x;
+	float y;
+	float z;
 	struct LinkedListNode *next;
 };
 struct CustomAVLTreeNode {
 	int height;
-	double coordinate_value;
+	float coordinate_value;
 	struct LinkedListNode *head_node;
 	struct CustomAVLTreeNode *left;
 	struct CustomAVLTreeNode *right;
@@ -22,12 +22,12 @@ public:
 
 	int Max(int a, int b);
 	int height(struct CustomAVLTreeNode *Node);
-	static struct CustomAVLTreeNode* newNode(double coordinate_value, double x, double y, double z);
+	static struct CustomAVLTreeNode* newNode(float coordinate_value, float x, float y, float z);
 	struct CustomAVLTreeNode* rightRotate(struct CustomAVLTreeNode *y);
 	struct CustomAVLTreeNode* leftRotate(struct CustomAVLTreeNode *x);
 	int getBalance(struct CustomAVLTreeNode *N);
-	struct CustomAVLTreeNode* insert(struct CustomAVLTreeNode* node, double coordinate_value, double x, double y, double z);
-	struct LinkedListNode* searchPoint(struct CustomAVLTreeNode* root, double x, double y, double z);
+	struct CustomAVLTreeNode* insert(struct CustomAVLTreeNode* node, float coordinate_value, float x, float y, float z);
+	struct LinkedListNode* searchPoint(struct CustomAVLTreeNode* root, float x, float y, float z);
 
 	void preOrder(struct CustomAVLTreeNode *root);
 };
